move generateSchedule to lib file, remove unused imports, clean up code
