+++ src/app/lib/placeholder-data.ts
@@ -1,5 +1,5 @@
-import { Availability, Task, generateSchedule } from "../generateSchedule";
-import { TaskTypes } from "./enum/tasktypes";
+import { Availability, Task } from "./generateSchedule";
+import { TaskTypes } from "./enum/taskTypes";
 export const availabilityOne: Availability = {
   startTime: new Date("2024-07-02T09:00:00"), // Wednesday, July 2nd at 9:00 AM
   endTime: new Date("2024-07-02T17:00:00"), // Wednesday, July 2nd at 5:00 PM
@@ -40,7 +40,3 @@ export const testTasks: Task[] = [
     duration: 30, // 30 minutes
   },
 ];
-
-// console.log(generateSchedule(availabilityOne, testTasks))
-// console.log("2nd Test")
-// console.log(generateSchedule(availabilityTwo, testTasks))
