+++ src/app/lib/tasks.ts
@@ -1,6 +1,7 @@
-export const unscheduledTasks = [];
+import { Task, Availability, ScheduledTasks } from "./generateSchedule";
+export const unscheduledTasks: (Task|ScheduledTasks)[] = [];
 
-export const availabilityTime = {
+export const availabilityTime : Availability = {
     startTime: null,
     endTime: null
 }
